// src/app/api/types.ts
export interface Advocate {
    city: string;
    degree: 'MD';
    firstName: string;
    lastName: string;
    phoneNumber: number
    specialties: string[];//    ['Relationship Issues (family, friends, couple, etc)', 'Trauma & PTSD', 'Personality disorders', 'Personal growth']
    yearsOfExperience: number;
}